ENG-729: consolidate PrintInvoice/PrintInvoices — extract InvoicePrintBody, InvoiceBillTo, InvoiceTitle
- Extract shared per-invoice rendering into InvoicePrintBody component
- Extract bill-to+QR header into InvoiceBillTo component
- Extract invoice title/date header into InvoiceTitle component
- Move getUnitComponentsByType, getApplicableTaxColumns, getBaseComponent helpers to InvoicePrintBody
- PrintInvoice.tsx reduced from 714 to 105 LOC
- PrintInvoices.tsx reduced from 819 to 145 LOC
- Add Playwright render-smoke spec for single and batch print routes
